Fixed compilation on GCC 3.X: Added (int) cast to forkexec() in libposix/unistd.cpp.
git-svn-id: https://freenos.googlecode.com/svn/trunk@95 824000e8-63cf-11de-a6e1-e33d94ba4d19

+++ lib/libposix/unistd.cpp
@@ -113,5 +113,5 @@ int forkexec(const char *path, const char *argv[])
     errno = msg.result;
     
     /* All done. */
-    return errno == ESUCCESS ? msg.number : -1;
+    return errno == ESUCCESS ? (int) msg.number : -1;
 }
